+++ tomcat/src/main/java/com/techcourse/handler/HelloHandler.java
@@ -1,22 +1,24 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.handler;
 
+import org.apache.coyote.http11.AbstractHandler;
+import org.apache.coyote.http11.ForwardResult;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpStatus;
 
 import java.net.URI;
 
-public class GetRegisterHandler extends AbstractHandler {
+public class HelloHandler extends AbstractHandler {
 
     @Override
     public boolean canHandle(HttpRequest httpRequest) {
         URI uri = httpRequest.getUri();
         String path = uri.getPath();
 
-        return "/register".equals(path) && httpRequest.getMethod().isGet();
+        return "/".equals(path);
     }
 
     @Override
     protected ForwardResult forward(HttpRequest httpRequest) {
-        return new ForwardResult("register.html", HttpStatus.OK);
+        return new ForwardResult("hello.html", HttpStatus.OK);
     }
 }
