+++ tomcat/src/main/java/com/techcourse/handler/NotFoundHandler.java
@@ -1,5 +1,7 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.handler;
 
+import org.apache.coyote.http11.AbstractHandler;
+import org.apache.coyote.http11.ForwardResult;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpStatus;
