+++ tomcat/src/test/java/com/techcourse/handler/GetLoginHandlerTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.handler;
 
 import org.apache.coyote.http11.Header;
 import org.apache.coyote.http11.HttpRequest;
