+++ tomcat/src/test/java/com/techcourse/handler/HelloHandlerTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.handler;
 
 import org.apache.coyote.http11.Header;
 import org.apache.coyote.http11.HttpRequest;
@@ -10,24 +10,24 @@ import java.util.Collections;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class PostRegisterHandlerTest {
+class HelloHandlerTest {
 
     @Test
-    @DisplayName("회원가입 관련 POST 요청을 처리할 수 있다.")
+    @DisplayName("루트 경로 요청을 처리할 수 있다.")
     void canHandle() {
-        PostRegisterHandler postRegisterHandler = new PostRegisterHandler();
+        HelloHandler helloHandler = new HelloHandler();
 
-        boolean result = postRegisterHandler.canHandle(createHttpRequest("POST /register HTTP/1.1"));
+        boolean result = helloHandler.canHandle(createHttpRequest("GET / HTTP/1.1"));
 
         assertThat(result).isTrue();
     }
 
     @Test
-    @DisplayName("회원가입 관련 POST 요청이 아니라면 처리할 수 없다.")
+    @DisplayName("루트 경로가 아닌 요청은 처리할 수 없다.")
     void cantHandle() {
-        PostRegisterHandler postRegisterHandler = new PostRegisterHandler();
+        HelloHandler helloHandler = new HelloHandler();
 
-        boolean result = postRegisterHandler.canHandle(createHttpRequest("GET /register HTTP/1.1"));
+        boolean result = helloHandler.canHandle(createHttpRequest("GET /login HTTP/1.1"));
 
         assertThat(result).isFalse();
     }
